+++ CHANGELOG.md
@@ -30,6 +30,10 @@
 
 ## Unreleased
 
+### Removed
+
+- Dropped support of Python 3.5.
+
 ## 1.6.0 - 2020-09-05
 
 ### Added